moved test_v8..test_v14 behind a test flag
