+++ synthesizer/Cargo.toml
@@ -49,6 +49,7 @@ serial = [
 ]
 setup = [ ]
 test = [ "snarkvm-console/test", "snarkvm-ledger-store/test", "snarkvm-synthesizer-process/test" ]
+test-old-consensus-versions = [ "test" ]
 timer = [ "aleo-std/timer" ]
 wasm = [
   "snarkvm-console/wasm",
