+++ synthesizer/src/vm/tests/mod.rs
@@ -13,24 +13,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(feature = "test")]
+// Tests specific to older consensus versions, not executed in most flows by default.
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v8;
 
-#[cfg(feature = "test")]
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v9;
 
-#[cfg(feature = "test")]
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v10;
 
-#[cfg(feature = "test")]
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v11;
 
-#[cfg(feature = "test")]
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v13;
 
-#[cfg(feature = "test")]
+#[cfg(feature = "test-old-consensus-versions")]
 mod test_v14;
 
+// Tests specific to recent consensus versions, executed in more flows than the older ones by
+// default.
 #[cfg(feature = "test")]
 mod test_v15;
 
@@ -45,3 +48,4 @@ mod test_v19;
 
 #[cfg(feature = "test")]
 use super::*;
+use crate::vm::test_helpers::*;
