+++ synthesizer/src/vm/tests/test_v15/mod.rs
@@ -15,10 +15,13 @@
 
 // Tests for the record-existence check.
 mod record_existence;
+
 // Tests on the input/output behaviour of closures and related functionality.
 mod closure_records;
+
 // Tests on the use of `commit_*_raw` instruction variants.
 mod commit_raw;
+
 // Additional test for cost estimation without a private key.
 mod cost_for_call;
 
@@ -40,5 +43,3 @@ use console::{
 
 use snarkvm_synthesizer_program::Program;
 use snarkvm_utilities::TestRng;
-
-use super::test_v14::add_and_test_with_costs;
