+++ synthesizer/src/vm/tests/test_v18/mod.rs
@@ -26,7 +26,6 @@ mod blockwide_synthesis_limit;
 
 use super::*;
 
-use super::test_v14::add_and_test_with_costs;
 use crate::vm::test_helpers::*;
 use console::account::ViewKey;
 use snarkvm_ledger_narwhal_subdag::test_helpers::subdag_with_cert_count;
