+++ synthesizer/src/vm/tests/test_v19/mod.rs
@@ -17,6 +17,3 @@
 mod translated_type_checks;
 
 use super::*;
-
-use super::test_v14::add_and_test_with_costs;
-use crate::vm::test_helpers::*;
